Improve chessbase control over game result
To support a wider variety of rules concerning game-terminating conditions
(such as mates and repeats), some aspects of the chessbase module are
put under user control:
1) Make number of repeats configurable:
A new parameter cbMaxRepeats is added that can be optionally set
in the game definition, to indicate on the how-manieth occurrence
of a position the game should be terminated. Default would be 3rd
occurrence (as before).
2) Count number of consecutive checks:
The variable this.check now is used as a counter rather than a boolean,
to keep track of the number of consecutive turns the current side to
move has been in check. (This is upward compatible with the old usage.)
A new variable this.oppCheck keeps track of the same for the other
player.
3) Allow custom evaluation of perpetuals:
When the chess base evaluation detects the maximum number of repetitions
is reached, it now calls a function cbPerpEval(this,aGame) for determining
the game result (which the function should return), if the game-specific
model has defined one. This way games with complex repetition rules
can take the result determination 'in their own hands'. If the function
is not defined the result is still determined in the old way, based on
the parameter cbOnPerpetual.
  When cbPerpEval returns undefined instead of a game result, the game
will continue. This is useful for implementing games where the ban on
repetition depends on the move that causes the repeat (e.g. to exempt
check evasions).
4) Allow custom evaluation of checkmates:
If the game-specific model has defined a function Model.Game.cbMateEval,
this function will be called when a checkmate is detected, and the
normal game result will be multiplied by what the function return,
to provide the game result. This can be used in variants that require
additional conditions to be fulfilled (which the function can then test).
E.g. in Shogi it can be used to invert the result after a Pawn-drop mate.
5) Make loss of all royals lose:
The chess base model applies the checking rule, and normally would not
allow a piece with isKing property to be captured. For some of the locust
moves this is difficult to enforce, though. As a backup it now also tests
for absence of any royals (as a free side effect of the kingMoved[]
determination), and flags that as a loss in games where the initial
position did contain royals (as seen by the traching of their location
in king[]).