Improve chessbase animation of lame oblique moves
Usually oblique moves are direct leaps, and would be animated by making
the piece jump by specifying a height in cbMoveMidZ. Non-jumping oblique
moves, such as those of the Xiangqi Horse or the Grant-Acedrex Gryphon,
can have the straight-line path to their destination run through occupied
squares, which looks rather ugly.
  To cure this cbMidMoveZ can now request the move to be animated purely
along diagonals and orthogonals. For this it has to return a negative
'hop height' h (i.e. the difference between the average origin and
destination height and the returned Z value). The animation routine
calculates the point where the rays over which to move intersect.
h = -1 then requests to move along the orthogonal first, h = -2
to do the diagonal first.
  The function View.Board.cbAnimate now gets the duration of a move
animation passed as an (optional) argument, rather than hard-coding
it as 600 (ms?). The default value is still 600, for compatibility
with existing code. But this patch makes it possible for variant-specific
overrides to manipulate the animation speed, which is for instance
desirable when animating bent multi-captures.