Provide better infra-structure for custom chess moves
Move generation in the chessbase infra-structure is extended by defining
three new flags. All of these would enable moves that would otherwise be
rejected to be pushed on a stack for special moves, which a custom
wrapper for the standard move generation can afterwards examine, to see
if they should give rise to pushing more moves on the regular move stack.
These move types are (1) moves to an empty square, (2) moves to a square
occupied by friend or (3) occupied by foe. Type (1) can e.g. be useful for
implementing Checkers-like capture, later testing if the jumped-over square
was occupied by a piece of the right player. Type (3) can be used for
identifying activation of own pieces (as in Knight-relay Chess), and (2)
for the first leg of a multi-capture (e.g. the Chu-Shogi Lion) or
(in combination with a null step) just to locate pieces that need completely
separate move generation.
  Candidate moves put on the 'specials' move stack will contain the
move descriptor that gave rise to their generation, as an 'x:' field.
This move descriptor contains all flags that were defined for this move
step when defining the corresponding move graph. This allows the use of
bits to which the move generator attaches no meaning as user-defined flags.
These could then indicate which type of post-processing the candidate moves
require, (e.g. locust capture on the square itself, on a jumped-over square,
or induction of new moves in friendly pieces), in variants that have
different kinds of special moves.
  The square number in the lower 16 bits of move.x is XORed with that
of the previously visited square, so that the latter can be derived by
special-move handlers that would want to know is by XORing with move.t.
  Special moves can be forced to contribute to the threatGraph by
setting FLAG_THREAT in their target specification, in addition to
FLAG_SPECIAL and/or FLAG_SPECIAL_CAPTURE. This is useful for moves that
can always capture to their destination, but which have to be generated
as specials because they have (or can have) side effects, or because
they might be duplicats (such as a jump to a distant square on a slider
path), which we might want to suppress.