New variant: Werewolf Chess
This variant makes use of the new multi-leg infra-structure, as the
Werewolf is a potentially double-capturing piece. Its jumps are defined
with both SPECIAL and SPECIAL_CAPTURE flags set, so they would always be
pushed as candidate moves. In addition they have the THREAT flag set.
This will not generate any moves, but forces inclusion of the jumps
in the threatGraph as if they were normal (i.e. unconditional) captures.
So that checks by a Werewolf will never be missed, and it will also
participate in static exchanges on a square it must reach by jumping.
 Wrappers around Model.Board.ApplyMove and Model.Board.cbQuickApply take
care of removal of any locust victim (indicated by the .via and .kill
fields of the move).