New fairy pieces: sized and deformed
Several new pieces are added to the 'fairy' set, generated from existing
ones by deforming them. For some this is just to give them a size that
is compatible with the King height of the set; these are named 'fr-proper-X',
with X = queen, marshal, cardinal, elephant or crowned Rook. These are
intended as replacements, and use the same 2d sprite as the originals.
Other pieces are meant as new types, and have their own 2d sprite.:
vertically sized rook and bishop (fr-small-rook, fr-small-bishop) for
range-limited versions, enhanced knights (fr-nightrider, fr-ferz-knight,
fr-wazir-knight, fr-zebra), a crowned Bishop (fr-crowned-bishop), a rook
with a tunnel running through it (fr-gate'), and a non-royal king (fr-man).
 The .js geometries are placed in the same directory as the piece from
which they were derived, with which they usually also share the material
texture maps.
  A 3d Hoplit image is added to the 'fairy' set, in the same directory as the
Pawn (with which it shares the normalmap). The 2d image is a normal pawn.
  Two kinds of pawns with a head the shape of a cube are added,
rotated 45-degrees w.r.t. each other. These can for instance be
used for insignificant pieces like Ferz and Wazir, after which they
are named in the fairy set (fr-ferz and fr-wazir).
  An Emperor is added as a deformed and upsized King (fr-emperor). It still
uses the same 2d image as a normal King.
  A demagnified version of the Eagle, with a rectified beak, is added to
the fairy set, under the name fr-birdie. It uses the same 2d image as the
original Eagle. It can be used to represend less impressive birds.
  Scirocco was referring to this piece as fr-proper-elephant, and its .js
file was in the fairy/elephant directory, but it was not mentioned in
fairy-set-view.js.
  The pieces that were added to the source for 2d images now also have
individual PNG files for use in rule-description HTML files and such.
  Since a merge has introduced two other pieces named fr-crowned-rook
and fr-crowned-bishop, the original pieces for this now go under the name
fr-proper-crowned-rook and fr-saint.
  A 200x120 PNG image of each finished piece with its 2d and 3d
representation is added to res/rules/fairy, for use in rule descriptions.
Also of existing pieces for which these did not exist yet.