New variant: Elven Chess
This 10x10 variant uses the multi-move infra-structure to implement the
moves of the Wizard (a Chu-Shogi Lion). Moves to occupied adjacent squares
are pushed as candidates on the specials stack. Afterwards all second legs
are generated for moves that hit a foe. (Moves that hit a friend are
rejected.) The restrictions on Wizard capture are implemented by detecting
WxW in cbQuickApply, and making the check test aware of that, so it can
treat the Lion as royal. Likewise it considers capture of an iron Lion
always illegal.