Allow chess base model to track multiple royals
Rather than tracking the location of just one royal, (with isKing set to
'true'), in this.kings[1] and [-1], the base model can now track multiple
royals with different isKing=N, in this.kings[N] or [-N]. CopyFrom makes
sure all the used elements in kings[] are copied.
  The value of royals is no longer overruled to 100, but set to their
definition, and only defaults to 100 when it was not defined.
  When cbGetAttackers is now called for the purpose of the check test,
it gets isKing passed to it as 100. Variant-specific wrappers can use
this to distinguish check testing (likely dependent on all royals) from
calls for static-exchange evaluation, which want to know captures on a
specific (expendable) royal.