New variant: Spartan Chess
This variant features multiple royals. The implementation makes use of
the new tracking mechanism supported by the chess base model, but this
is not sufficient: the base model does not update the location of the
royals when they are captured. So ApplyMove and cbQuickApply still need
wrappers.
 When cbGetAttackers is used for the purpose of check detection it tests
attacks on both Kings to impolement the duple-check rule.
  Static Exchange Evaluation was also a problem, because how the Kings
are sorted in the capture order, as well as how much score they bring in
when captured, depends on whether you have one or two. This is now solved
by calling cbStaticExchangeEval through a wrapper that first manipulates
the King values in the pTypes table depending on presence of a spare.
It does not take account of whether the spare King is under attack, so that
trading the other would put you in check, but the exchange evaluation in
general ignores whether the captures put you in check or not.