Let general handling of repeats track loop length
Rather than just storing the number of visits to each position,
HandleRepeat() now also stores the move number (obtained as length of
the mPlayedMoves array) where the position first occurred (packed as
16*moveNr + occurrenceCount). The function GetRepeatOccurrence() can be
used with a second, non-zero argument to return the number of half-moves
done since the first occurrence of the position. This infrastructure can
be used for scoring perpetuals in games that forbid certain actions (such
as checking in Xiangqi) on every move of a repeat loop, to know which moves
should be scrutinized.