Fragile repair of 50-move rule in chess base
For declaring 50-move draws no account was taken of the fact that not
only captures, but also Pawn pushes reset the counter. This is tricky
to fix in just the chessbase model, because it requires one to know
which pieces are 'Pawns'. (Or, more generally, which piece types should
reset the counter on moving.) Even in games with only one kind of Pawn
the latter are often represented by multiple internal types: white and
black, moved or virgin.
  For future game implementations a configurable parameter cbPawnTypes
is provided, which would cause all types < cbPawnTypes to reset the
noCaptCount on moving. This thus requires Pawns to be specified as the
lowest piece types in the game definition, as seems to be habitually
done in existing implementations of chess variants.
  The default value for cbPawnTypes is set in a tricky way: it is
assumed that the type defined first is a Pawn, and that alternative
Pawn types follow it. It thus sets cbPawnTypes to the specified type
number of the first type (in definition order) that has another value
for the 'abbrev' property. This would be satisfactory when all types
to be used for Pawns are defined first, with the lowest numbers.
When sticking to that rule, only games that have truly different pieces
that should reset the counter need to explicitly define cbPawnTypes.
  This seems to work for most existing chess-variant definitions, so that
their individual model files don't have to be modified. Even where it
doesn't work it would just replace one type of 50-move bug by another.