Implement ban on perpetual checking in Xiangqi
Scoring of 3-fold repetition in Xiangqi (repeater loses) had no
similarity whatsoever with the official rules (draw unless King or
other piece would be incessantly harrassed). This is now partly fixed:
perpetual checking (= repeat through all checking moves) is now detected,
and the player doing it will be set to lose (no matter who caused the
repetition). In other cases the game will be declared a draw.
(So perpetual chasing of non-royals, which could still be a tie breaker
in that case, is not yet detected.)