New piece set: oriental Shogi tiles
The module 'shogi-set-view.js' contains a set of kanji tiles for the
pieces of regular Shogi. The 2d images are 'bare kanji', printed
upside down for gote. Hand counters are provided with the 2d set.
The 'aspect' names all start with 'sh-', e.g. sh-gold, sh-dragon,
sh-tokin, sh-promoted-knight. The Gyokusho is provided as sh-jade.

KNOWN PROBLEM: the 2d view doesn't behave as one would desire when
viewed as Player B!