New variant: Shogi
Shogi is implemented in the framework of the chessbase/drop sub-model.
Because the board is uncheckered, the view supplies grid lines on top
of the standard 3d grid board. This is ugly on the dark squares used in
the holdings area, but in the latest design of the drop view thes are
totally covered by the counters anyway.
  It also uses the infra-structure of cbMateEval (to score Pawn-drop
mates) and cbPerpEval (to punish perpetual checking).