Fix repetition detection in chess base model
Repetitions could be overlooked in the chess base model, because
different copies of identical pieces (same type and color) were
considered different, so that swapping their location would not be
recognized as the same position. This because the position hash was
based on the index of the piece in the piece list, rather than on its
(colored) type. In normal chess such swapping is very unlikely, but
in variants where captured pieces can be dropped back it happens quite
often.
  This fix bases the position hash on type, color and location only,
ignoring the piece index. The old API function based on index and
square now accesses the piece of the given index to get its type and
color, and uses 3*type+color instead of the index as first selector
of the key seed, in 'board' mode. This makes the use of 'type' mode
superfluous, so that this is changed in a no-op.
  A new API for key updating is provided, through three functions:
Game.Model.bKey, tKey and wKey, which have to be called for obtaining
a 'board' key, 'type' key or 'who' key, respectively. These functions
return a value with which the caller has to XOR this this.zSign key himself.
The take a piece object as argument, and use the properties they need from
it (the position piece.p and eithr the index piece.i or the type p.t).
Whe the definition contains a field zobrist: "old" these functions return
the key change that would mimic the old zobrist.update functions. This
is needed in games that have an opening book, which is organized by zSign.
Other games should use the new zonbist scheme, which detects repetitons
correctly.
  Metamachy, Zanzibar-S, Chess960 and LeyChessAlpha were accessing the old
Zobrist interface directly, for shuffling their setup. Hence the new hashing
scheme had broken them. Wildebeest Chess, Modern Chess and Shafran Chess
accessed it for special moves (non-standard castling and Bishopp swaps).
These are now all patched to use the new API.
  Classic Chess and Xiangqi had an opening book, and have gotten a
zobrist: "old", property in their cbDefine object.

Zanzibar-S: adapt to new Zobrist API (and fix)

The old API was accessed for setting up the position from the game's
model file. The old code was buggy anywhay, as it used the location of
the white Rhino for the key of the black one.