New sub-model: support common fairy and multi-leg chess moves
This sub-model, when used in combination with base-model.js, offers three
kinds of new support:
1) Some Graph functions for common fairy pieces: Camel, Zebra, Alibaba,
Champion, Wizard, Griffon, Rhino. The latter two are bent sliders, and
a general Graph function cbSkiSlide is added to facilitate adding this
kind of piece, where the first step can have other flags than the remaining
part of the path. A value of -1 here means the first square is skipped
(i.e not part of the path). A larger negative value for the initial flags
can be used to shift the square where the 2nd leg of the trajectory will
start, so that pieces like D-then-B can be implemented as well.
For iflags = -1 the corner is still skipped (which is good for Ski-sliders
and the Grant-Acedrex Unicorno), but for -2 two steps along the initial
vector will be made before the trajectory bends, etc. The corner then gets
the same flags as the rest of the path. The bending angle must be passed
in steps of 45 degees; for positive values bending will be in both directions,
for negative value only in the specified one.
2) Anti-trade rules: piece definitions can now have a field antiTrade,
which specifies the number of an 'anti-trade group'. Pieces are not allowed
to capture a protected piece in the same trade group. If the group number
is larger than 100 they are not even allowed to capture an unprotected
piece in the same group. Furthermore, when the group number is negative,
there also applies a 'counter-strike' rule, and you cannot capture a piece
in the group if one was captured on the immediately preceding move.
  In fact an even number N and the following N+1 specify the same anti-trade
group, but the piece with the even antiTrade can be captured always, and is
only restricted in capturing the piece with the odd antiTrade value.
3) Multi-leg moves: some new mode flags are defined, for hit & run capture,
rifle capture, and checker capture. These all will cause pushing of the
indicated move on the special-moves stack, but the standard move generation
is extended to recognize those there, and finish the processing.
In particular, hit & run moves will indicate a capture that must be followed
by another step in arbitrary direction. Rifle capture will create a
back-and-forth move to the target, and checker capture (to be used only with
leaps of length 2) capture a jumped-over enemy. The multi-leg moves only
work when the multi-leg-view.js file is included in the view build script.
  The second step of hit & run by default is a King move that can both
capture and go to empty, but can be specified different by assigning a move
graph to Model.Game.neighbors. The mode flags specified here, will indicate
whether the extra move away from the capture square can be to an empty square
(FLAG_MOVE), capture an enemy (FLAG_CAPTURE) or back to the move's
starting square (FLAG_RIFLE).
  A new FLAG_BURN is added to the move-modality constants in the FLAG_FAIRY
group. When used with FLAG_HITRUN it causes all enemies adjacent to the
destination to be removed as an automatic side effect of the move. The
extension and effect of this 'burn zone' can be controlled by assigning
a move graph to Model.Game.burnZone, and by setting FLAG_CAPTURE_SELF
as mode in this graph the piece can also be made to burn friendly pieces.
  Another enhancement is that InitGame now calls an optional routine
'extraInit' in the Game object, with geometry as argument. This allows
sub-models to do some additional initialization after the geometry has
been initialized.