Support flying capture in the chessbase module
Flying capture is a special case of screen capture, where a piece can
fly over multiple screens. In the first place base-model.js was improved
by allowing screen capture and normal capture to be specified in the same
move. The move generation for screen capture, as well as cbGetAttackers()
was changed to allow flying capture for pieces that have a new field
'ranking' specified (and non-zero) in their piece-type definition.
Pieces with a higher ranking will block the captures of pieces with a
lower (non-zero) ranking, which otherwise would be able to jump arbitrary
many pieces to make the capture. Pieces of equal ranking only block each
other's captures when that ranking is odd.