Fix entry and animation of non-standard castlings
The animation in base-model.js now pays attention to a modification
of the rook destination specified in move.t. The move entry routine
(which has a modified copy in multi-leg-view.js) is changed such that
only the castling specified in the table will require clicking of the
rook as destination; if there are other castlings with the same rook
these must be specified by clicking on the kings destination square.
This can support flexible castling by defining the one-step castling
in the table, and generate other king/rook destinations as custom moves.