New variant: Makromachy
A 14x14 chess variant in the chessbase module, with 2x56 pieces.
It relies heavily on the sub-model fairy-move-model, as it involves
hit & run capture. The Eagle, Raven and Bat are represented by the
flying sliders, and the new Champion is used.
  The new default cbMoveMidZ can handle all Makromachy pieces
satisfactorily, so no dedicated replacement is provided in the view file.
  In the rule description he pieces are organized as a table, so image and
description appear side by side. Both a 2d and 3d 600x600 image are added.