Support piece definitions from FEN in fairy-move-model.js
A routine Model.Game.cbPiecesfFromFEN(geometry,fen,step) is added.
It will return an object that defines a chess variant with the indicated
pieces and setup, which in principle can be directly returned by cbDefine.
  The piece IDs that can be used in the FEN are not only the orthodox
pieces PNBRQK, but also some commonly used fairy pieces: The Wizard (W)
and Champion (O) from Omega Chess, the Archbishop (A) and Marshall (M)
from Capablanca Chess, the Camel (C), (Modern) Elephant (E) and Zebra (Z),
The Xiangqi Cannon (X) and its diagonal counterpart the Vao (V), the bent
sliders Griffon (G) and Manticore (U), the Crowned Rook (D) and Bishop (H),
and the Amazon (T) and Lion (L).
  Pawns are assumed to be e.p. capable, and Rook are assumed to be
castle pieces. A new routine to generate Pawn graphs is added:
cbFlexiPawnGraph(geometry, direction, pawnRank, maxPush).
This generates a position-dependent move, where on all ranks up to and
including the specified pawnRank the Pawn can be pushed to halfway the
board, or by maxPush, whichever is lower. By default pawnRank and maxPush
are equal to the board height, leading to a Pawn that can always be pushed
to halfway the board, no matter where it is and whether it has moved yet.
  By specifying pawnRank as the rank where the (white) Pawns start, Pawns
that have moved will be no longer on it, and thus lose the initial push.
  The routine cbPiecesFromFEN now accepts pawnRank and maxPush as 3rd and
4th argument too; here the default for the pawnRank is the rank where the
(first encountered) white Pawn starts.
  Apart from the pieceTypes, the returned object also specifies promotion
and castling. The default King move increases with board width. It also has
a method setCastling(kstep, partnerName) that can be used to alter the
castlings for moving the King by another amount, or castling with a piece
different from Rook.
  For promotion it assumes one can pick any non-royal piece on the last rank.
This can be altered by assigning an array with piece types to the object's
property promoChoice, a large number to promoZone.
  The returned  object is also equiped with methods addPiece(typeDefinition),
setProperty(name, property, value) and addMoves(name, graph) to tweak the
definitions of the automatically generated types, or add pieces of types
that are not provided automatically. It also contains a method
setValues(object,property), which as its first argument accepts an object
with property names as the piece-type abbreviations, that as value have the
newly desired values of the property for that type given by the second
argument. The default property is the piece 'value'. Like
p.setValues({P: 0.8, N:2.5, B:4, R:6}) to specify new values for Pawn,
Knight, Bishop and Rook.

The routine to generate Pawn graphs in fairy-move-model.js is now
cbFlexiPawnGraph(geometry, direction, pawnRank, maxPush).
This generates a position-dependent move, where on all ranks up to and
including the specified pawnRank the Pawn can be pushed to halfway the
board, or by maxPush, whichever is lower. By default pawnRank and maxPush
are equal to the board height, leading to a Pawn that can always be pushed
to halfway the board, no matter where it is and whether it has moved yet.
  By specifying pawnRank as the rank where the (white) Pawns start, Pawns
that have moved will be no longer on it, and thus lose the initial push.
  The routine cbPiecesFromFEN now accepts pawnRank and maxPush as 3rd and
4th argumet too; here the default for the pawnRank is the rank where the
(first encountered) white Pawn starts.

There also is a new function cbLionGraph(geometry, confine).