Improve default cbMoveMidZ for grid board
Rather than only jumping on oblique moves, cbMoveMidZ now examines the
move graph for the moved piece, to see if the destination is reached
in a single step, and only jumps when it is (and is not adjacent).
It also jumps for a destination further along a path that is flagged
as screen capture, when the move is a capture.
  For oblique slides the routine now investigates whether the diagonal
or orthogonal leg should be animated first. It then explicitly requests
bent sliding for oblique slides, and selects the appropriate order of the
legs by examining the first destination on the path.
  If the first leap in the path to the destination is not to an adjacent
square, a jumping move animation is requested.