Add function to streamline move-graph generation
A function Model.Game.cbSymmetricGraph(geometry,spec,confine) was added,
which can generate a mixture of moves with different ranges and flags.
Each number in the array specs is interpreted either as a new flags
setting pertaining to the moves that follow, or the description of an
8-fold-symmetric move set. The units of this number indicate the x-step
and the decades the y-stap; any higher-order digits the range. A preceding
minus sign indicates infinite range, and the default range is 1.
So an Archbishop would require [21,-11], where 21 is the [2,1] leap of
the Knight, and -11 the [1,1] slide of the Bishop. (The flags setting
starts as FLAG_MOVE|FLAG_CAPTURE.)