New variant: Minjiku Shogi
A 10x10 dropless version, inspired by Tenjiku Shogi.
A customGen routine is used, which converts special null moves (requested
through FLAG_CAPTURE_SELF on a [0,0] step) to all Knight moves if one of
the two adjacent intermediate squares is empty, and to all War Machine
moves if the direct slide is blocked, and one of the two diagonally
adjacent intermediates is empty.
  This needs to be complemented with the radial slides (which
could possibly have longer range) to make a complete area move.
  In the rule description the pieces are presented as a table, so the text
appears next to the images instead of above those. Both 2d and 3d 600x600
images are included.