Chessbase: fix lastMove handling
Since Move.c contains null for a non-capture and the piece index otherwise,
it is a frequent source of bugs that there also exists a piece with index 0.
The evaluation did not do the test for determining whether the move was a
capture (and cbStaticExchangeEval had to be invoked) properly.
  Board.lastMove is now initialized with a (recognizable) dummy Move
object rather than null; this only affects the initial board, and
simplifies CopyFrom. It also opens the possibility to make Board.lastMove
do double-duty as an indicator for the stage of a possible game prelude for
setting up an initial position, so that CopyFrom does not have to be
burdened with copying additional properties.