Chessbase: split up cbInitialPosition
The part to sort the piece list and set up the board is taken out of
cbInitialPosition, and put in a new routine cbPlacePieces, called by
the former. This makes it possible for other code to sort the list
again after a modification has been made (e.g. after promotiong some
pieces to other types).