New chess sub-model: prelude
There now are prelude-model.js and prelude-view.js files, which can be
included in a build for games where a player first has to select a setup
before proper playing can begin. The prelude has the form of a number
of popups, one of which will be shown during a turn. Each popup contains
a number of buttons labeled with piece icons, which can be clicked to
place the corresponding pieces on a set of predefined squares.
  For this to work the game-definition object must have a property
'prelude', which is an array, where each element corresponds to a turn
at the beginning of the game. A zero there indicates that turn will be
passed automatically, allowing the same player to do two things in a row,
or black to start the setup selection.
  Otherwise the element must be an object {setup:[...], squares:[...]}.
'setup' must be an array of strings, where each string indicates a number
of pieces by their ID (=(fen)abbrev property). These pieces will be placed
(in the given order) on the set of 'squares' given as an array of square
numbers.
  The setup strings can contain slashes, to force the icons of the
specified pieces to appear in multiple rows on the button, or spaces,
to indicate an icon-sized empty spot on the button. (This has no
consequence for the eventual placement of these pieces.) This can be
used to give the buttons a layout similar to the squares where the
pieces will be placed. E.g. "AB/CD" will place icons for pieces A-D
in a 2x2 pattern, A and B on the first row, leading to square buttons.
  For each setup string a button will appear in a 'selection panel',
and an optional property 'panelWidth' controls how many of these buttons
will be displayed on one row of the panel. The default for this tries
to make the panel as square as possible.
  If the computer should play the turn for which the panel is defined,
no panel will appear, and it will select one of the setups randomly.
  The object that defines the panel can have an optional property
'participants'. Its value can be an aray that exists outside the object;
after a selection is made this array will be filled with all non-royal,
non-pawn types that are then on the board. This could be used to update
the array returned by the 'promote' routine, when rules apply that you
can promote to any type originally present, and the selection will
eliminate some types.