Fairy-move-model: support partial graph merging
The method addMoves in the game-definition object returned by cbPiecesFromFEN,
which can be used to merge an extra move graph into the graph of the mentioned
piece type, now accepts an array of square numbers as a third parameter.
In that case only the moves starting from the thus specified squares will
be merged. This is useful for defining non-standard castlings, which can
happen only from the initial square of the castling piece.