Prelude: support persistent selections
The selection made through a prelude dialog can now be declared persistent
through a property persistent: true in the dialog definition. In that case
the same selection will be automatically applied in subsequent games. This
is useful when the selection is used as a second-level variant selection
within a group of variants, rather than a single variant with variable
initial position.
  The dialog definition can now also have an optional property 'castling',
which must contain an array with as many members as there are setups,
each containing a castling definition. This so variants with different
castling rules can be grouped.
  It also has optional properties 'participants' and 'blackParticipants',
which must refer to arrays.  After a selection these arrays will be filled
with the non-royal, non-pawn pieces that are present on teh board in this
new setup. This can be used to update promoChoice arrays used by a promote
routine. That black can be treated separately is for the benefit of variants
with asymmetric pieces.
  Finally there is a property custom, which canrefer to a function that
will be called after the selection, with the number of the selected setup
as parameter (as well as the current board and game objects).