Capablanca Chess: turn into prelude reference implementation
The new prelude sub-model is used to let the Capablanca Chess implementation
support an entire group of 10x8 variants. Since some of these have flexible
castling, a customGen routine had to be added to support that through the
fairy-move-model. In total 10 variants are supported, with different setups
and castling rules.