Chessbase: support flexible castling
The castle table in the game definition only allows one castling per
King/Rook pair, as this combination is used as index to find the spec.
In some variants the King can move a variable number of steps during
castling, though. This is now implemented by allowing a property 'extra'
in the spec, which must contain a number indicating how much further the
King is allowed to move.
  The tabulated castling must be entered by clicking on the Rook as
destination, the extra castlings by clicking the King destination.
If the value of extra is negative, its sign is flipped, but the castling
in the table is then suppressed. (So that all castlings that are allowed
can be entered by clicking the King destination.)