Chessbase: add function for defining castlings in fairy-piece-model
A new function cbCastlingDef returns the object that needs to go in the
castle property of a game definition. It must be called with ten square
numbers: the king origin, destination on the 'short' side, rook on the
short side, destination on the 'long' side and rook on that side. And then
the same for black. The short side will be notated as O-O; it doesn't
matter whther it is on the left or right, or if it is actually long or
short.