Chessbase: allow pieces to start off board
When the start location in the game definition is specified as -1,
the piece will be created in the piece list, but not be put on he board.
This is needed in variants where the number of pieces can increase
compared to the initial position, e.g. due to piece-creating moves.