Team-Mate Chess: Implement Brute promotion
The rule addition that a Pawn can promote to a pair of Brutes is implemented,
through wrappers of ApplyMove and cbQuickApply, which add the second Brute
on the from-square of a Brute promotion. A move diagram for the Brute was
added to the rule description.
  The promotion panel had to be shrunk a bit to fit the extra option in
the 2d representation, and it was made to rely on the default cbMoveMidZ.
  The game-specific evaluation is changed to exempt the KPK end-game from
being declared 'insufficient material' draw, as the other 2-vs-1 end-games
are. It was also enhanced to pay attention to color binding, penalizing
imbalance on the square shades, and even declaring draw if a bare King is
facing only pieces that are bound to the same shade.
  The evaluation now makes use of the same evaluation parameters as
Classic Chess; before it used the Shatranj parameters by mistake.