Scirocco: let evaluation score promotability
The dedicated evaluate function now estimates which fraction of each
player's pieces will eventually protode (i.e. after trading what was
needed to undermine the enemy defense of the promotion zone), and
how much you will gain by that. For leapers this value is discounted
for a large part, depending on their distance from the zone.