Chessbase: pass pieceValue to dedicated evaluate
The evalValues that are passed to the game's dedicate evaluate function
only contain the difference of the pieceValues and pieceValueRatio, not
the total value of the individual white and black armies. Unfortunately
the latter cannot be reconstructed from teh former, if they are equal.
But for a serious evaluation these are essential, expecially in variants
where all pieces can promote (to score the latent promotion value, which
is very game-phase- dependent).
  For now this is solved by passing the calculated army strength in the
pieceValue object of Evaluate() to the dedicated evaluate as 4th argument.
(A better design would be to put those in evalValues instead of pieceValue
and pieceValueRatio, and let the dedicated evaluate calculate these by
itself if it wants to use them, but it would be hard to do this in a
backward-compatible way.)